refactor(access-control): move listAllUsers query into @op/common service
Extract the platform-admin user-listing logic (cursor pagination, email
search, sentinel exclusion, whereWeWork shaping) out of the tRPC router
into packages/common/src/services/platform/listAllUsers.ts, matching how
other list routes delegate to a canonical service. The router keeps only
the rate-limit + platform-admin middleware, input/output schemas, and the
userEncoder boundary.
