+++ packages/common/src/services/platform/index.ts
@@ -1,1 +1,2 @@
+export * from './listAllUsers';
 export * from './stats';
